# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/invitations:
    post:
      operationId: InvitationsController_createInvitation
      summary: Create invitation for collaborator
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationDto'
      responses:
        '201':
          description: Invitation created successfully
        '400':
          description: Bad request
        '403':
          description: Forbidden
      tags:
        - Invitations
      security:
        - gym-id: []
        - bearer: []
    get:
      operationId: InvitationsController_getGymInvitations
      summary: Get pending invitations for gym
      parameters:
        - name: gymId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of pending invitations
      tags:
        - Invitations
      security:
        - gym-id: []
        - bearer: []
  /api/v1/invitations/{token}/accept:
    post:
      operationId: InvitationsController_acceptInvitation
      summary: Accept invitation
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInvitationDto'
      responses:
        '200':
          description: Invitation accepted successfully
        '400':
          description: Invalid or expired invitation
        '404':
          description: Invitation not found
      tags:
        - Invitations
  /api/v1/invitations/{id}/cancel:
    put:
      operationId: InvitationsController_cancelInvitation
      summary: Cancel pending invitation
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Invitation cancelled successfully
        '400':
          description: Cannot cancel this invitation
        '404':
          description: Invitation not found
      tags:
        - Invitations
      security:
        - bearer: []
...
  /api/v1/gyms:
    post:
      operationId: GymsController_createGym
      summary: Create a new gym
      parameters:
        - name: organizationId
          required: true
          in: query
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGymDto'
      responses:
        '201':
          description: Gym created successfully
        '400':
          description: Bad request
        '403':
          description: Forbidden - Gym limit reached
      tags:
        - Gyms
      security:
        - bearer: []
    get:
      operationId: GymsController_getOrganizationGyms
      summary: Get gyms for organization
      parameters:
        - name: organizationId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of gyms
      tags:
        - Gyms
      security:
        - bearer: []
  /api/v1/gyms/{id}:
    get:
      operationId: GymsController_getGym
      summary: Get gym details
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Gym details
        '404':
          description: Gym not found
      tags:
        - Gyms
      security:
        - bearer: []
        - gym-id: []
    put:
      operationId: GymsController_updateGym
      summary: Update gym details
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGymDto'
      responses:
        '200':
          description: Gym updated successfully
        '404':
          description: Gym not found
      tags:
        - Gyms
      security:
        - bearer: []
        - gym-id: []
  /api/v1/gyms/{id}/stats:
    get:
      operationId: GymsController_getGymStats
      summary: Get gym statistics
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Gym statistics
        '404':
          description: Gym not found
      tags:
        - Gyms
      security:
        - bearer: []
        - gym-id: []
  /api/v1/gyms/{id}/toggle-status:
    put:
      operationId: GymsController_toggleGymStatus
      summary: Toggle gym active status
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Gym status toggled successfully
        '404':
          description: Gym not found
      tags:
        - Gyms
      security:
        - bearer: []
...
  /api/v1/evaluations:
    post:
      operationId: EvaluationsController_createEvaluation
      summary: Create a new evaluation for a client
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEvaluationDto'
      responses:
        '201':
          description: Evaluation created successfully
        '400':
          description: Bad request
      tags:
        - Evaluations
      security:
        - gym-id: []
        - bearer: []
  /api/v1/evaluations/{id}:
    get:
      operationId: EvaluationsController_getEvaluation
      summary: Get evaluation details
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Evaluation details
        '404':
          description: Evaluation not found
      tags:
        - Evaluations
      security:
        - gym-id: []
        - bearer: []
    put:
      operationId: EvaluationsController_updateEvaluation
      summary: Update evaluation
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEvaluationDto'
      responses:
        '200':
          description: Evaluation updated successfully
        '404':
          description: Evaluation not found
      tags:
        - Evaluations
      security:
        - gym-id: []
        - bearer: []
    delete:
      operationId: EvaluationsController_deleteEvaluation
      summary: Delete evaluation
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Evaluation deleted successfully
        '404':
          description: Evaluation not found
      tags:
        - Evaluations
      security:
        - gym-id: []
        - bearer: []
  /api/v1/evaluations/client/{clientId}:
    get:
      operationId: EvaluationsController_getClientEvaluations
      summary: Get client evaluation history
      parameters:
        - name: clientId
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Client evaluation history with evolution metrics
      tags:
        - Evaluations
      security:
        - gym-id: []
        - bearer: []
  /api/v1/evaluations/gym/stats:
    get:
      operationId: EvaluationsController_getGymEvaluationStats
      summary: Get gym evaluation statistics
      parameters: []
      responses:
        '200':
          description: Gym evaluation statistics
      tags:
        - Evaluations
      security:
        - gym-id: []
        - bearer: []
  /api/v1/evaluations/{id}/report:
    get:
      operationId: EvaluationsController_generateEvaluationReport
      summary: Generate evaluation report
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Evaluation report data
      tags:
        - Evaluations
      security:
        - gym-id: []
        - bearer: []
...
info:
  title: GymSpace API
  description: The GymSpace Gym Management System API
  version: '1.0'
  contact:
    name: GymSpace API Support
    email: support@gymspace.com
  license:
    name: MIT
...
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.gymspace.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    gym-id:
      type: apiKey
      in: header
      name: X-Gym-Id
  schemas:
...
    CreateInvitationDto:
      type: object
      properties:
        email:
          type: string
          example: john.trainer@example.com
        gymId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        roleId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174001
      required:
        - email
        - gymId
        - roleId
    AcceptInvitationDto:
      type: object
      properties:
        name:
          type: string
          example: John Trainer
        phone:
          type: string
          example: '+1234567890'
        password:
          type: string
          example: SecurePassword123!
      required:
        - name
        - phone
        - password
...
    CreateGymDto:
      type: object
      properties:
        name:
          type: string
          example: PowerFit Downtown
        address:
          type: string
          example: 123 Main Street
        city:
          type: string
          example: New York
        state:
          type: string
          example: NY
        postalCode:
          type: string
          example: '10001'
        phone:
          type: string
          example: '+1234567890'
        email:
          type: string
          example: info@powerfit.com
        openingTime:
          type: string
          example: '08:00'
        closingTime:
          type: string
          example: '22:00'
        capacity:
          type: number
          example: 150
        amenities:
          type: object
          example:
            hasParking: true
            hasShowers: true
            hasLockers: true
        settings:
          type: object
          example:
            logo: https://example.com/logo.png
            primaryColor: '#FF0000'
      required:
        - name
    UpdateGymDto:
      type: object
      properties:
        name:
          type: string
          example: PowerFit Downtown
        address:
          type: string
          example: 123 Main Street
        city:
          type: string
          example: New York
        state:
          type: string
          example: NY
        postalCode:
          type: string
          example: '10001'
        phone:
          type: string
          example: '+1234567890'
        email:
          type: string
          example: info@powerfit.com
        openingTime:
          type: string
          example: '08:00'
        closingTime:
          type: string
          example: '22:00'
        capacity:
          type: number
          example: 150
        amenities:
          type: object
          example:
            hasParking: true
            hasShowers: true
            hasLockers: true
        settings:
          type: object
          example:
            logo: https://example.com/logo.png
            primaryColor: '#FF0000'
...
    CreateEvaluationDto:
      type: object
      properties:
        gymClientId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        weight:
          type: number
          example: 75.5
          description: Weight in kg
        height:
          type: number
          example: 175
          description: Height in cm
        bodyFatPercentage:
          type: number
          example: 25
          description: Body fat percentage
        muscleMassPercentage:
          type: number
          example: 40
          description: Muscle mass percentage
        measurements:
          type: object
          example:
            chest: 100
            waist: 85
            hips: 95
            leftArm: 30
            rightArm: 30
            leftThigh: 55
            rightThigh: 55
          description: Body measurements in cm
        healthMetrics:
          type: object
          example:
            bloodPressure: 120/80
            restingHeartRate: 65
            vo2Max: 45
          description: Health metrics
        performanceMetrics:
          type: object
          example:
            benchPress: 80
            squat: 100
            deadlift: 120
          description: Performance metrics in kg
        notes:
          type: string
          example: Client has improved significantly
        goals:
          type: string
          example: Build muscle mass and reduce body fat
        recommendations:
          type: string
          example: Increase protein intake and maintain training frequency
      required:
        - gymClientId
        - weight
        - height
    UpdateEvaluationDto:
      type: object
      properties:
        gymClientId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        weight:
          type: number
          example: 75.5
          description: Weight in kg
        height:
          type: number
          example: 175
          description: Height in cm
        bodyFatPercentage:
          type: number
          example: 25
          description: Body fat percentage
        muscleMassPercentage:
          type: number
          example: 40
          description: Muscle mass percentage
        measurements:
          type: object
          example:
            chest: 100
            waist: 85
            hips: 95
            leftArm: 30
            rightArm: 30
            leftThigh: 55
            rightThigh: 55
          description: Body measurements in cm
        healthMetrics:
          type: object
          example:
            bloodPressure: 120/80
            restingHeartRate: 65
            vo2Max: 45
          description: Health metrics
        performanceMetrics:
          type: object
          example:
            benchPress: 80
            squat: 100
            deadlift: 120
          description: Performance metrics in kg
        notes:
          type: string
          example: Client has improved significantly
        goals:
          type: string
          example: Build muscle mass and reduce body fat
        recommendations:
          type: string
          example: Increase protein intake and maintain training frequency
...